Allowing customisation of the generated widths classes
We're using HAML as our templating engine, and the default fractional system is invalid syntax. There are ugly workarounds; but it seems to me that it ought to be possible for end users to override the generated grid classes. I recall that previously inuit supported words or fractions via a variable; which I'd be in favour of. I've got a sample implementation if you'd like to see it in a seperate PR.

Here I've taken the approach of putting the / separator into a variable than can be overridden, and then pulling the fractions out of the list themselves which would allow an end user to set $inuit-fractions to be words and everything should just work.

Sample usage:
$inuit-fractions: one two three four five;
$inuit-space-separator: "-of-";

(generates .u-one-of-two)