[refs #346] Enhance ratio/crop object
The ratio modifier classes for the ratio/crop object used to be
defined by a ‘simple’ Sass list, whose entries were used for the
required calculation as well as the class name. This involved
a lot of Sass magic and also ended up quite unflexible concerning
class naming for ‘special’ ratios like the golden-ratio or custom
ones.

Now it is possible to determine the ratio as well as the according
class name, so you can name your modifier class e.g.
`o-ratio--golden` or `o-crop--foo`.